+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-16 03:07:28
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 2a6a84a103a0b083f57751883a9fb0b9801d3d2f
+
+## [ForwardWidgets-2kuai] - 2025-08-16 03:07:28
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 4ff60f5b522a8a6e40e1360f4439f957ad950c2c
+
+## [ForwardWidgets-pack1r] - 2025-08-16 03:07:28
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: dde1da3793d648e92d70e3e4c2bc568c50dafb9e
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-15 03:14:22
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main